+++ app/Helpers/LocaleHelper.php
@@ -91,20 +91,32 @@ class LocaleHelper
      */
     public static function getLocaleList()
     {
-        $locales = collect([]);
-        foreach (config('lang-detector.languages') as $lang) {
-            $name = trans('settings.locale_'.$lang);
-            if ($name == 'settings.locale_'.$lang) {
-                // The name of the new language is not already set, even in english
-                $name = $lang;
-            }
-            $locales->push([
+        return collect(config('lang-detector.languages'))->map(function ($lang) {
+            return [
                 'lang' => $lang,
-                'name' => $name,
-            ]);
+                'name' => self::getLocaleName($lang),
+                'name-orig' => self::getLocaleName($lang, $lang),
+            ];
+        });
+    }
+
+    /**
+     * Get the name of one language.
+     *
+     * @param string $lang
+     * @param string $locale
+     *
+     * @return string
+     */
+    private static function getLocaleName($lang, $locale = null) : string
+    {
+        $name = trans('settings.locale_'.$lang, [], $locale);
+        if ($name == 'settings.locale_'.$lang) {
+            // The name of the new language is not already set, even in english
+            $name = $lang;
         }
 
-        return $locales->sortByCollator('name');
+        return $name;
     }
 
     /**