+++ tests/Unit/Models/ContactTest.php
@@ -381,26 +381,26 @@ class ContactTest extends FeatureTestCase
         $contact = factory(Contact::class)->create();
 
         $activity1 = factory(Activity::class)->create([
-            'date_it_happened' => '2015-10-29 10:10:10',
+            'happened_at' => '2015-10-29',
             'account_id' => $contact->account_id,
         ]);
         $contact->activities()->attach($activity1, ['account_id' => $contact->account_id]);
 
         $activity2 = factory(Activity::class)->create([
-            'date_it_happened' => '2010-10-29 10:10:10',
+            'happened_at' => '2010-10-29',
             'account_id' => $contact->account_id,
         ]);
         $contact->activities()->attach($activity2, ['account_id' => $contact->account_id]);
 
         $activity3 = factory(Activity::class)->create([
-            'date_it_happened' => '1981-10-29 10:10:10',
+            'happened_at' => '1981-10-29',
             'account_id' => $contact->account_id,
         ]);
         $contact->activities()->attach($activity3, ['account_id' => $contact->account_id]);
 
         $this->assertEquals(
-            '2015-10-29 10:10:10',
-            $contact->getLastActivityDate()
+            '2015-10-29',
+            $contact->getLastActivityDate()->format('Y-m-d')
         );
     }
 
@@ -409,14 +409,14 @@ class ContactTest extends FeatureTestCase
         $contact = factory(Contact::class)->create();
 
         $activity1 = factory(Activity::class)->create([
-            'date_it_happened' => '2015-10-29 10:10:10',
+            'happened_at' => '2015-10-29',
             'account_id' => $contact->account_id,
         ]);
         $contact->activities()->attach($activity1, ['account_id' => $contact->account_id]);
 
         $this->assertEquals(
-            '2015-10-29 10:10:10',
-            $contact->getLastActivityDate()
+            '2015-10-29',
+            $contact->getLastActivityDate()->format('Y-m-d')
         );
     }
 
@@ -462,25 +462,6 @@ class ContactTest extends FeatureTestCase
         $this->assertEquals(
             7,
             strlen($contact->default_avatar_color)
-        );
-    }
-
-    public function testUpdateFoodPreferencesSetsNullIfEmptyValueGiven()
-    {
-        $contact = factory(Contact::class)->create();
-        $contact->updateFoodPreferences('');
-
-        $this->assertNull($contact->food_preferences);
-    }
-
-    public function testUpdateFoodPreferencesEncryptsTheValue()
-    {
-        $contact = factory(Contact::class)->make();
-        $contact->updateFoodPreferences('Some value');
-
-        $this->assertEquals(
-            'Some value',
-            $contact->food_preferences
         );
     }