+++ CHANGELOG.md
@@ -7,6 +7,13 @@
 
 ### Enhancements:
 
+* Add foreign keys to all tables
+* Add English (UK) locale
+* Add API methods to destroy and store documents
+* Add API methods to manage photos and avatars
+* Add emotions and participants to activities
+* Activities are now added inline
+* Enable API web navigation
 * Enhance UI of API's Settings to add comprehension and documentation
 * Improve trim string middleware to not trim password text
 * Upgrade to Laravel 6.5
@@ -18,6 +25,7 @@
 
 ### Fixes:
 
+* Fix birthdate selection UX
 * Fix OAuth login process with WebAuthn activated
 * Fix journal entry edit
 * Fix register in case country is not detected from ip address